pyproject.toml
[project]
name = "personanexus"
version = "0.1.0"
description = ""
authors = [
    {name = "Naufal Hadi Darmawan",email = "128771220+NaufalHD12@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "django-friendship (>=1.9.6,<2.0.0)",
    "django-allauth (>=65.9.0,<66.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "django-anymail[mailgun] (>=13.0,<14.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "django-browser-reload (>=1.18.0,<2.0.0)",
    "django-htmx (>=1.23.0,<2.0.0)",
    "django-tinymce (>=4.1.0,<5.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-tailwind (>=1.0.3,<2.0.0)"
]

[tool.poetry]
packages = [{include = "personanexus", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
